Refactor files to ensure passing tets
