+++ app/models.py
@@ -1,5 +1,5 @@
-from pewpewpew import db
-class Game(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    player_name = db.Column(db.String(64))
-    score = db.Column(db.Integer)
+# from pewpewpew import db
+# class Game(db.Model):
+#     id = db.Column(db.Integer, primary_key=True)
+#     player_name = db.Column(db.String(64))
+#     score = db.Column(db.Integer)
